j07 + j05 fix

+++ PiscineC2018/j05/ex07/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: flklein <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/09/02 00:00:43 by flklein           #+#    #+#             */
-/*   Updated: 2018/09/02 00:01:27 by flklein          ###   ########.fr       */
+/*   Updated: 2018/09/04 21:48:15 by flklein          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@ int		ft_strncmp(char *s1, char *s2, unsigned int n)
 	int		i;
 
 	i = 0;
-	while (i < n && s1[i] && s2[i] && s1[i] == s2[i])
+	while (i < (int)n && s1[i] && s2[i] && s1[i] == s2[i])
 		i++;
 	return (s1[i] - s2[i]);
 }
